Add files via upload
This project is a basic chatbot using Python. It talks with the user using fixed replies. The user types something, and the bot gives a reply based on that input.

The chatbot asks the user to type something.

If the user types "hello", the bot replies with "Hi!".

If the user types "how are you", it replies with "I'm fine, thanks!".

If the user types "bye", the chatbot says "Goodbye!" and ends.

If the user types anything else, the bot says it doesn’t understand.


It runs in a loop and keeps replying until the user says "bye".